Use Identifier.Sanitize on parameters and values
There's a risk of sql injection and other weirdness here, since
the parameters and values are user supplied. Of course, the user
in this case is the database admin, but still better to clean
the input up. We quote these values as identifiers (double quotes)
because postgres accepts them as such.