Rework timeout goroutine with more synchronization to avoid races
Implementing timeout via cancel is trickier than it seems, with a
lot of potential race conditions. The cancel might fire before
the statement fires, or after it completes, leaving us in
and indeterminate state. Add a bunch of synchronization, with cancel
retry after 100ms if the main thread hasn't continued after a
cancel was fired.